Replace urllib with requests poster isn't available for Python3, so this is the only way to get multipart for file upload
Heavily based on pull request #3 by vaab

Works for basic queries and login, untested on uploads and HTTP Auth

This also eliminates the ability to store the cookie data to a file.
It's not actually a requirement to do this, but that was a lot of
old, complicated code for a feature that (IMO) isn't very important.